Port to 1.10 and Dogecoin implementation
There are many changes that I cannot remember, but I have refactored all
code to change the name of this plugin from "Satoshis" to "Dogecoins"
and have added libdohj as well as updated BitcoinJ to the latest version
to change the wallet system to use the Dogecoin network. The dogecoin
blockchain appears to download and JSON 'database' is implemented.